requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.3
django-cors-headers==3.13.0
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
Faker==18.3.2
idna==3.4
joblib==1.2.0
numpy==1.24.2
pandas==2.0.0
Pillow==9.4.0
psycopg2-binary==2.9.6
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.2
scikit-surprise==1.1.3
scipy==1.10.1
six==1.16.0
sqlparse==0.4.3
toposort==1.10
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
